% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_perc.R
\name{as_perc}
\alias{as_perc}
\title{Format a proportion as a percentage}
\usage{
as_perc(p, digits = 0)
}
\arguments{
\item{p}{A length-1 numeric to be interpreted as a proportion}

\item{digits}{Number of digits to round percentage to (default to 0)}
}
\value{
Returns a string to report a percentage to the specified number of
  digits.
}
\description{
\code{as_perc} formats a proportion as a percentage to print in an RMarkdown
document
}
\details{
Simply multiplies \code{p} by 100 and affixes a percent sign to the end after
rounding.
}
\examples{
as_perc(0.2345)

as_perc(0.000234)

}
