% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewerRender.R
\name{tldr}
\alias{tldr}
\title{Instantiate tldR manpage.}
\usage{
tldr(name)
}
\arguments{
\item{name}{A string containing the slash-saparated package/function query.}
}
\value{
An object of class 'tldr'
}
\description{
Instantiate tldR manpage.
}
\examples{
t <- tldr("base/paste", render=FALSE)
tldr("base/paste")
}
\keyword{tldr}
