% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_compare.R
\name{cont_compare}
\alias{cont_compare}
\title{Compare a numerical variable across levels of a categorical variable}
\usage{
cont_compare(y, grp, plot = c("density", "boxplot", "none"))
}
\arguments{
\item{y}{A numerical variable}

\item{grp}{A categorical variable}

\item{plot}{Type of plot to produce}
}
\value{
Returns a list including (1) group-wise summary statistics, (2) ANOVA decomposition, (3) eta-squared effect size, and (4) ggplot2 object, if requested.
}
\description{
Deprecated. Use `\code{\link{num_compare}}` instead.
}
