% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_compare.R
\name{cont_compare}
\alias{cont_compare}
\title{Compare a numerical variable across levels of a categorical variable}
\usage{
cont_compare(y, grp, plot = c("density", "boxplot", "none"))
}
\description{
Deprecated. Use \code{\link{`num_compare`}} instead.
}
