% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tldr-package.R, R/tldr.R
\name{tldr}
\alias{tldr}
\alias{tldr-package}
\title{tldr: Short-form documentation in the console}
\usage{
tldr(topic)
}
\arguments{
\item{topic}{A name or character string specifying documentation topic}
}
\description{
A package that implements quick "tldr"-style documentation in the R console,
providing the \code{tldr()} function for users and the roxygen2 extension
\code{tldr-roclet} for package developers.

Similar to \verb{?} and \code{help()}
}
\examples{
tldr(tldr)
tldr("tldr")
}
