% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tldr.R
\name{tldr}
\alias{tldr}
\title{Short form, tl;dr style documentation in the console}
\usage{
tldr(topic)
}
\arguments{
\item{topic}{A name or character string specifying documentation goal}
}
\value{

}
\description{
Similar to \verb{?} and \code{help()}
}
\examples{
tldr(tldr)
tldr("tldr")
tldr(palmerpenguins::penguins)

}
