% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-extension.R
\name{tldr_roclet}
\alias{tldr_roclet}
\alias{roxy_tag_rd_tldr}
\alias{roxy_tag_rd.roxy_tag_paramtldr}
\alias{roxy_tag_rd_tldr.roxy_tag_paramtldr}
\alias{roxy_tag_rd.roxy_tag_exampletldr}
\alias{roxy_tag_rd_tldr.roxy_tag_exampletldr}
\alias{roxy_tag_rd.roxy_tag_formattldr}
\alias{roxy_tag_rd_tldr.roxy_tag_formattldr}
\alias{roxy_tag_rd.roxy_tag_ignoretldr}
\alias{roxy_tag_rd_tldr.roxy_tag_ignoretldr}
\alias{block_to_rd_tldr}
\alias{block_to_rd_tldr.roxy_block}
\alias{roxy_tag_parse.roxy_tag_paramtldr}
\alias{format.rd_section_paramtldr}
\alias{roxy_tag_parse.roxy_tag_exampletldr}
\alias{format.rd_section_exampletldr}
\alias{roxy_tag_parse.roxy_tag_formattldr}
\alias{format.rd_section_formattldr}
\alias{roxy_tag_parse.roxy_tag_ignoretldr}
\alias{format.rd_section_ignoretldr}
\alias{roclet_process.roclet_tldr}
\alias{roclet_clean.roclet_tldr}
\alias{roclet_output.roclet_tldr}
\title{Roxygen2 extensions}
\usage{
roxy_tag_rd_tldr(x, base_path, env)

\method{roxy_tag_rd}{roxy_tag_paramtldr}(x, base_path, env)

\method{roxy_tag_rd_tldr}{roxy_tag_paramtldr}(x, base_path, env)

\method{roxy_tag_rd}{roxy_tag_exampletldr}(x, base_path, env)

\method{roxy_tag_rd_tldr}{roxy_tag_exampletldr}(x, base_path, env)

\method{roxy_tag_rd}{roxy_tag_formattldr}(x, base_path, env)

\method{roxy_tag_rd_tldr}{roxy_tag_formattldr}(x, base_path, env)

\method{roxy_tag_rd}{roxy_tag_ignoretldr}(x, base_path, env)

\method{roxy_tag_rd_tldr}{roxy_tag_ignoretldr}(x, base_path, env)

block_to_rd_tldr(block, base_path, env)

\method{block_to_rd_tldr}{roxy_block}(block, base_path, env)

\method{roxy_tag_parse}{roxy_tag_paramtldr}(x)

\method{format}{rd_section_paramtldr}(x, ...)

\method{roxy_tag_parse}{roxy_tag_exampletldr}(x)

\method{format}{rd_section_exampletldr}(x, ...)

\method{roxy_tag_parse}{roxy_tag_formattldr}(x)

\method{format}{rd_section_formattldr}(x, ...)

\method{roxy_tag_parse}{roxy_tag_ignoretldr}(x)

\method{format}{rd_section_ignoretldr}(x, ...)

tldr_roclet()

\method{roclet_process}{roclet_tldr}(x, blocks, env, base_path)

\method{roclet_clean}{roclet_tldr}(x, base_path)

\method{roclet_output}{roclet_tldr}(x, results, base_path, ..., is_first = FALSE)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}

\item{block}{Roxygen block to parse to rd}

\item{...}{Additional arguments}

\item{blocks}{List of roxygen blocks}

\item{results}{Rd code resulting from parsed roxygen blocks}

\item{is_first}{Has roclet_output written files before}
}
\description{
roxygen2 extension code facilitating documenting via roxygen2::roxygenize(roclets = c("tldr_roclet"))
}
